+++ src/pybamm/models/full_battery_models/lead_acid/base_lead_acid_model.py
@@ -64,7 +64,7 @@ class BaseModel(pybamm.BaseBatteryModel):
     def set_soc_variables(self):
         """Set variables relating to the state of charge."""
         # State of Charge defined as function of electrolyte concentration
-        z = pybamm.standard_spatial_vars.z
+        z = pybamm.SpatialVariable("current collector")
         soc = (
             pybamm.Integral(
                 self.variables["X-averaged electrolyte concentration [mol.m-3]"]