+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 - Removed standard spatial variables. `SpatialVariable` objects should be created explicitly instead. Also, `SpatialVariable` objects no longer take in a "name", and the coordinate system is now specified in the geometry instead of in the spatial variable
 - Refactored the geometry classes to use a custom `Domain` class that can be used to create 1D, 2D, and 3D domains, including tabs.
 - Removed the "macroscale" spatial method helper
--
+- `var_pts` keys should now be domain names (e.g. `"negative particle"`) instead of spatial variable names (e.g. `"r_n"`)
 
 # [v24.11.0](https://github.com/pybamm-team/PyBaMM/tree/v24.11.0) - 2024-11-20